Abstract out a PythonDocTestOrCodeBlockParser
This is used in Markdown variants.
